requirements.txt
streamlit
pandas
plotly
pypdf
python-docx
openpyxl
